+++ web/apps/photos/src/components/FixCreationTime.tsx
@@ -1,4 +1,5 @@
 import log from "@/base/log";
+import { EnteFile } from "@/media/file";
 import {
     decryptPublicMagicMetadata,
     fileCreationPhotoDate,
@@ -9,7 +10,6 @@ import { FileType } from "@/media/file-type";
 import { PhotoDateTimePicker } from "@/new/photos/components/PhotoDateTimePicker";
 import downloadManager from "@/new/photos/services/download";
 import { extractExifDates } from "@/new/photos/services/exif";
-import { EnteFile } from "@/new/photos/types/file";
 import { fileLogID } from "@/new/photos/utils/file";
 import { ensure } from "@/utils/ensure";
 import DialogBox from "@ente/shared/components/DialogBox/";
