+++ web/apps/photos/src/components/PhotoFrame.tsx
@@ -1,9 +1,9 @@
 import log from "@/base/log";
+import type { LivePhotoSourceURL, SourceURLs } from "@/media/file";
+import { EnteFile } from "@/media/file";
 import { FileType } from "@/media/file-type";
 import type { GalleryBarMode } from "@/new/photos/components/gallery/BarImpl";
 import DownloadManager from "@/new/photos/services/download";
-import type { LivePhotoSourceURL, SourceURLs } from "@/new/photos/types/file";
-import { EnteFile } from "@/new/photos/types/file";
 import { PHOTOS_PAGES } from "@ente/shared/constants/pages";
 import { CustomError } from "@ente/shared/error";
 import useMemoSingleThreaded from "@ente/shared/hooks/useMemoSingleThreaded";
