+++ web/apps/photos/src/services/publicCollectionService.ts
@@ -5,8 +5,8 @@ import type {
     Collection,
     CollectionPublicMagicMetadata,
 } from "@/media/collection";
+import { EncryptedEnteFile, EnteFile } from "@/media/file";
 import { sortFiles } from "@/new/photos/services/files";
-import { EncryptedEnteFile, EnteFile } from "@/new/photos/types/file";
 import { mergeMetadata } from "@/new/photos/utils/file";
 import { CustomError, parseSharingErrorCodes } from "@ente/shared/error";
 import HTTPService from "@ente/shared/network/HTTPService";
