+++ web/apps/photos/src/utils/collection.ts
@@ -8,9 +8,9 @@ import {
     CollectionType,
     SUB_TYPE,
 } from "@/media/collection";
+import { EnteFile } from "@/media/file";
 import { ItemVisibility } from "@/media/file-metadata";
 import { getAllLocalFiles, getLocalFiles } from "@/new/photos/services/files";
-import { EnteFile } from "@/new/photos/types/file";
 import { safeDirectoryName } from "@/new/photos/utils/native-fs";
 import { CustomError } from "@ente/shared/error";
 import { LS_KEYS, getData } from "@ente/shared/storage/localStorage";
