+++ web/apps/photos/src/utils/magicMetadata/index.ts
@@ -1,7 +1,7 @@
 import { sharedCryptoWorker } from "@/base/crypto";
 import type { Collection } from "@/media/collection";
+import { EnteFile, MagicMetadataCore } from "@/media/file";
 import { ItemVisibility } from "@/media/file-metadata";
-import { EnteFile, MagicMetadataCore } from "@/new/photos/types/file";
 
 export function isArchivedFile(item: EnteFile): boolean {
     if (!item || !item.magicMetadata || !item.magicMetadata.data) {
