+++ web/packages/new/photos/services/files.ts
@@ -1,7 +1,7 @@
 import { blobCache } from "@/base/blob-cache";
+import { type EnteFile, type Trash } from "@/media/file";
 import { FileType } from "@/media/file-type";
 import localForage from "@ente/shared/storage/localForage";
-import { type EnteFile, type Trash } from "../types/file";
 import { mergeMetadata } from "../utils/file";
 
 const FILES_TABLE = "files";
