+++ web/packages/new/photos/services/ml/index.ts
@@ -9,8 +9,8 @@ import log from "@/base/log";
 import { masterKeyFromSession } from "@/base/session-store";
 import type { Electron } from "@/base/types/ipc";
 import { ComlinkWorker } from "@/base/worker/comlink-worker";
+import type { EnteFile } from "@/media/file";
 import { FileType } from "@/media/file-type";
-import type { EnteFile } from "@/new/photos/types/file";
 import { ensure } from "@/utils/ensure";
 import { throttled } from "@/utils/promise";
 import { proxy, transfer } from "comlink";
