+++ web/packages/new/photos/services/search/types.ts
@@ -5,9 +5,9 @@
 
 import type { Location } from "@/base/types";
 import type { Collection } from "@/media/collection";
+import type { EnteFile } from "@/media/file";
 import { FileType } from "@/media/file-type";
 import type { Person } from "@/new/photos/services/ml/people";
-import type { EnteFile } from "@/new/photos/types/file";
 import type { LocationTag } from "../user-entity";
 
 /**
