+++ web/packages/new/photos/services/search/worker.ts
@@ -1,8 +1,8 @@
 import { HTTPError } from "@/base/http";
 import type { Location } from "@/base/types";
 import type { Collection } from "@/media/collection";
+import type { EnteFile } from "@/media/file";
 import { fileCreationPhotoDate, fileLocation } from "@/media/file-metadata";
-import type { EnteFile } from "@/new/photos/types/file";
 import { ensure } from "@/utils/ensure";
 import { nullToUndefined } from "@/utils/transform";
 import { getPublicMagicMetadataSync } from "@ente/shared/file-metadata";
