+++ web/packages/new/photos/utils/file.ts
@@ -2,9 +2,9 @@ import { isDesktop } from "@/base/app";
 import log from "@/base/log";
 import { CustomErrorMessage } from "@/base/types/ipc";
 import { workerBridge } from "@/base/worker/worker-bridge";
+import type { EnteFile } from "@/media/file";
 import { needsJPEGConversion } from "@/media/formats";
 import { heicToJPEG } from "@/media/heic-convert";
-import type { EnteFile } from "../types/file";
 import { detectFileTypeInfo } from "./detect-type";
 
 /**
